refactor: consolidated all github buttons under a subdirectory
